Refactor ifw and hd driver codes to use the common helper methods from driver_utilities.py
